Refactored integration.py in jira to have a field builder for custom jira field stuff and issues to separate integration and issues related functionality.
