+++ src/sentry/integrations/jira/integration.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 
 import logging
-import six
+
 
 from django.core.urlresolvers import reverse
 from django.utils.translation import ugettext as _
@@ -10,14 +10,15 @@ from sentry import features
 from sentry.integrations import (
     IntegrationInstallation, IntegrationFeatures, IntegrationProvider, IntegrationMetadata, FeatureDescription,
 )
-from sentry.integrations.exceptions import ApiUnauthorized, ApiError, IntegrationError, IntegrationFormError
-from sentry.integrations.issues import IssueSyncMixin
-from sentry.models import IntegrationExternalProject, Organization, OrganizationIntegration, User
+from sentry.integrations.exceptions import ApiError, IntegrationError
+from sentry.models import IntegrationExternalProject, Organization
 from sentry.utils.http import absolute_uri
 
 from .client import JiraApiClient, JiraCloud
+from .field_builder import JiraFieldBuilder
+from .issues import JiraIssueSync
 
-logger = logging.getLogger('sentry.integrations.jira')
+logger = logging.getLogger('sentry.integrations.jira')  # not used here anymore
 
 DESCRIPTION = """
 Connect your Sentry organization into one or more of your Jira cloud instances.
@@ -73,23 +74,8 @@ metadata = IntegrationMetadata(
     },
 )
 
-# hide sprint, epic link, parent and linked issues fields because they don't work
-# since sprint and epic link are "custom" we need to search for them by name
-HIDDEN_ISSUE_FIELDS = {
-    'keys': ['parent', 'issuelinks'],
-    'names': ['Sprint', 'Epic Link'],
-}
 
-# A list of common builtin custom field types for Jira for easy reference.
-JIRA_CUSTOM_FIELD_TYPES = {
-    'select': 'com.atlassian.jira.plugin.system.customfieldtypes:select',
-    'textarea': 'com.atlassian.jira.plugin.system.customfieldtypes:textarea',
-    'multiuserpicker': 'com.atlassian.jira.plugin.system.customfieldtypes:multiuserpicker',
-    'tempo_account': 'com.tempoplugin.tempo-accounts:accounts.customfield'
-}
-
-
-class JiraIntegration(IntegrationInstallation, IssueSyncMixin):
+class JiraIntegration(IntegrationInstallation, JiraFieldBuilder, JiraIssueSync):
     comment_key = 'sync_comments'
     outbound_status_key = 'sync_status_forward'
     inbound_status_key = 'sync_status_reverse'
@@ -236,21 +222,6 @@ class JiraIntegration(IntegrationInstallation, IssueSyncMixin):
 
         self.model.save()
 
-    def get_link_issue_config(self, group, **kwargs):
-        fields = super(JiraIntegration, self).get_link_issue_config(group, **kwargs)
-        org = group.organization
-        autocomplete_url = reverse(
-            'sentry-extensions-jira-search', args=[org.slug, self.model.id],
-        )
-        for field in fields:
-            if field['name'] == 'externalIssue':
-                field['url'] = autocomplete_url
-                field['type'] = 'select'
-        return fields
-
-    def get_issue_url(self, key, **kwargs):
-        return '%s/browse/%s' % (self.model.metadata['base_url'], key)
-
     def get_persisted_default_config_fields(self):
         return ['project', 'issuetype', 'priority']
 
@@ -278,41 +249,6 @@ class JiraIntegration(IntegrationInstallation, IssueSyncMixin):
             verify_ssl=True
         )
 
-    def get_issue(self, issue_id, **kwargs):
-        client = self.get_client()
-        issue = client.get_issue(issue_id)
-        return {
-            'key': issue_id,
-            'title': issue['fields']['summary'],
-            'description': issue['fields']['description'],
-        }
-
-    def create_comment(self, issue_id, user_id, group_note):
-        # https://jira.atlassian.com/secure/WikiRendererHelpAction.jspa?section=texteffects
-        comment = group_note.data['text']
-        quoted_comment = self.create_comment_attribution(user_id, comment)
-        return self.get_client().create_comment(issue_id, quoted_comment)
-
-    def create_comment_attribution(self, user_id, comment_text):
-        user = User.objects.get(id=user_id)
-        attribution = '%s wrote:\n\n' % user.name
-        return '%s{quote}%s{quote}' % (attribution, comment_text)
-
-    def update_comment(self, issue_id, user_id, group_note):
-        quoted_comment = self.create_comment_attribution(user_id, group_note.data['text'])
-        return self.get_client().update_comment(
-            issue_id, group_note.data['external_id'], quoted_comment,
-        )
-
-    def search_issues(self, query):
-        try:
-            return self.get_client().search_issues(query)
-        except ApiError as e:
-            self.raise_error(e)
-
-    def make_choices(self, x):
-        return [(y['id'], y['name'] if 'name' in y else y['value']) for y in x] if x else []
-
     def error_message_from_json(self, data):
         message = ''
         if data.get('errorMessages'):
@@ -337,389 +273,6 @@ class JiraIntegration(IntegrationInstallation, IssueSyncMixin):
         return reverse(
             'sentry-extensions-jira-search', args=[org_slug, self.model.id]
         )
-
-    def build_dynamic_field(self, group, field_meta):
-        """
-        Builds a field based on Jira's meta field information
-        """
-        schema = field_meta['schema']
-
-        # set up some defaults for form fields
-        fieldtype = 'text'
-        fkwargs = {
-            'label': field_meta['name'],
-            'required': field_meta['required'],
-        }
-        # override defaults based on field configuration
-        if (schema['type'] in ['securitylevel', 'priority']
-                or schema.get('custom') == JIRA_CUSTOM_FIELD_TYPES['select']):
-            fieldtype = 'select'
-            fkwargs['choices'] = self.make_choices(field_meta.get('allowedValues'))
-        elif field_meta.get('autoCompleteUrl') and \
-                (schema.get('items') == 'user' or schema['type'] == 'user'):
-            fieldtype = 'select'
-            fkwargs['url'] = self.search_url(group.organization.slug)
-            fkwargs['choices'] = []
-        elif schema['type'] in ['timetracking']:
-            # TODO: Implement timetracking (currently unsupported alltogether)
-            return None
-        elif schema.get('items') in ['worklog', 'attachment']:
-            # TODO: Implement worklogs and attachments someday
-            return None
-        elif schema['type'] == 'array' and schema['items'] != 'string':
-            fieldtype = 'select'
-            fkwargs.update(
-                {
-                    'multiple': True,
-                    'choices': self.make_choices(field_meta.get('allowedValues')),
-                    'default': ''
-                }
-            )
-
-        # break this out, since multiple field types could additionally
-        # be configured to use a custom property instead of a default.
-        if schema.get('custom'):
-            if schema['custom'] == JIRA_CUSTOM_FIELD_TYPES['textarea']:
-                fieldtype = 'textarea'
-
-        fkwargs['type'] = fieldtype
-        return fkwargs
-
-    def get_issue_type_meta(self, issue_type, meta):
-        issue_types = meta['issuetypes']
-        issue_type_meta = None
-        if issue_type:
-            matching_type = [t for t in issue_types if t['id'] == issue_type]
-            issue_type_meta = matching_type[0] if len(matching_type) > 0 else None
-
-        # still no issue type? just use the first one.
-        if not issue_type_meta:
-            issue_type_meta = issue_types[0]
-
-        return issue_type_meta
-
-    def get_create_issue_config(self, group, **kwargs):
-        kwargs['link_referrer'] = 'jira_integration'
-        fields = super(JiraIntegration, self).get_create_issue_config(group, **kwargs)
-        params = kwargs.get('params', {})
-
-        defaults = self.get_project_defaults(group.project_id)
-        project_id = params.get('project', defaults.get('project'))
-
-        client = self.get_client()
-
-        # If we don't have a jira project selected, fetch the first project
-        # This avoids a potentially very expensive API call to fetch issue
-        # create configuration for *all* projects.
-        jira_projects = client.get_projects_list()
-        if not project_id and len(jira_projects):
-            project_id = jira_projects[0]['id']
-
-        try:
-            meta = client.get_create_meta_for_project(project_id)
-        except ApiUnauthorized:
-            raise IntegrationError(
-                'Jira returned: Unauthorized. '
-                'Please check your configuration settings.'
-            )
-        except ApiError as exc:
-            logger.info(
-                'jira.error-fetching-issue-config',
-                extra={
-                    'integration_id': self.model.id,
-                    'organization_id': group.organization.id,
-                    'error': exc.message,
-                }
-            )
-            raise IntegrationError(
-                'There was an error communicating with the Jira API. '
-                'Please try again or contact support.'
-            )
-        if not meta:
-            raise IntegrationError(
-                'No projects were found in Jira. Check the permissions for projects.'
-            )
-
-        # check if the issuetype was passed as a parameter
-        issue_type = params.get('issuetype', defaults.get('issuetype'))
-        issue_type_meta = self.get_issue_type_meta(issue_type, meta)
-        issue_type_choices = self.make_choices(meta['issuetypes'])
-
-        # make sure default issue type is actually
-        # one that is allowed for project
-        if issue_type:
-            if not any((c for c in issue_type_choices if c[0] == issue_type)):
-                issue_type = issue_type_meta['id']
-
-        fields = [
-            {
-                'name': 'project',
-                'label': 'Jira Project',
-                'choices': [(p['id'], p['key']) for p in jira_projects],
-                'default': meta['id'],
-                'type': 'select',
-                'updatesForm': True,
-            }
-        ] + fields + [
-            {
-                'name': 'issuetype',
-                'label': 'Issue Type',
-                'default': issue_type or issue_type_meta['id'],
-                'type': 'select',
-                'choices': issue_type_choices,
-                'updatesForm': True,
-            }
-        ]
-
-        # title is renamed to summary before sending to Jira
-        standard_fields = [f['name'] for f in fields] + ['summary']
-
-        # TODO(jess): are we going to allow ignored fields?
-        # ignored_fields = (self.get_option('ignored_fields', group.project) or '').split(',')
-        ignored_fields = set(
-            k for k, v in six.iteritems(issue_type_meta['fields'])
-            if v['name'] in HIDDEN_ISSUE_FIELDS['names']
-        )
-        ignored_fields.update(HIDDEN_ISSUE_FIELDS['keys'])
-
-        # apply ordering to fields based on some known built-in Jira fields.
-        # otherwise weird ordering occurs.
-        anti_gravity = {"priority": -150, "fixVersions": -125, "components": -100, "security": -50}
-
-        dynamic_fields = issue_type_meta['fields'].keys()
-        dynamic_fields.sort(key=lambda f: anti_gravity.get(f) or 0)
-
-        # build up some dynamic fields based on required shit.
-        for field in dynamic_fields:
-            if field in standard_fields or field in [x.strip() for x in ignored_fields]:
-                # don't overwrite the fixed fields for the form.
-                continue
-            mb_field = self.build_dynamic_field(group, issue_type_meta['fields'][field])
-            if mb_field:
-                mb_field['name'] = field
-                fields.append(mb_field)
-
-        for field in fields:
-            if field['name'] == 'priority':
-                # whenever priorities are available, put the available ones in the list.
-                # allowedValues for some reason doesn't pass enough info.
-                field['choices'] = self.make_choices(client.get_priorities())
-                field['default'] = defaults.get('priority', '')
-            elif field['name'] == 'fixVersions':
-                field['choices'] = self.make_choices(client.get_versions(meta['key']))
-
-        return fields
-
-    def create_issue(self, data, **kwargs):
-        client = self.get_client()
-        cleaned_data = {}
-        # protect against mis-configured integration submitting a form without an
-        # issuetype assigned.
-        if not data.get('issuetype'):
-            raise IntegrationFormError({'issuetype': ['Issue type is required.']})
-
-        jira_project = data.get('project')
-        if not jira_project:
-            raise IntegrationFormError({'project': ['Jira project is required']})
-
-        meta = client.get_create_meta_for_project(jira_project)
-        if not meta:
-            raise IntegrationError('Could not fetch issue create configuration from Jira.')
-
-        issue_type_meta = self.get_issue_type_meta(data['issuetype'], meta)
-
-        fs = issue_type_meta['fields']
-        for field in fs.keys():
-            f = fs[field]
-            if field == 'description':
-                cleaned_data[field] = data[field]
-                continue
-            elif field == 'summary':
-                cleaned_data['summary'] = data['title']
-                continue
-            elif field == 'labels' and 'labels' in data:
-                labels = [
-                    label.strip()
-                    for label in data['labels'].split(',')
-                    if label.strip()
-                ]
-                cleaned_data['labels'] = labels
-                continue
-            if field in data.keys():
-                v = data.get(field)
-                if not v:
-                    continue
-
-                schema = f.get('schema')
-                if schema:
-                    if schema.get('type') == 'string' and not schema.get('custom'):
-                        cleaned_data[field] = v
-                        continue
-                    if schema['type'] == 'user' or schema.get('items') == 'user':
-                        v = {'name': v}
-                    elif schema.get('custom') == JIRA_CUSTOM_FIELD_TYPES.get('multiuserpicker'):
-                        # custom multi-picker
-                        v = [{'name': v}]
-                    elif schema['type'] == 'array' and schema.get('items') != 'string':
-                        v = [{'id': vx} for vx in v]
-                    elif schema['type'] == 'array' and schema.get('items') == 'string':
-                        v = [v]
-                    elif schema.get('custom') == JIRA_CUSTOM_FIELD_TYPES.get('textarea'):
-                        v = v
-                    elif (schema['type'] == 'number' or
-                          schema.get('custom') == JIRA_CUSTOM_FIELD_TYPES['tempo_account']):
-                        try:
-                            if '.' in v:
-                                v = float(v)
-                            else:
-                                v = int(v)
-                        except ValueError:
-                            pass
-                    elif (schema.get('type') != 'string'
-                            or (schema.get('items') and schema.get('items') != 'string')
-                            or schema.get('custom') == JIRA_CUSTOM_FIELD_TYPES.get('select')):
-                        v = {'id': v}
-                cleaned_data[field] = v
-
-        if not (isinstance(cleaned_data['issuetype'], dict) and 'id' in cleaned_data['issuetype']):
-            # something fishy is going on with this field, working on some Jira
-            # instances, and some not.
-            # testing against 5.1.5 and 5.1.4 does not convert (perhaps is no longer included
-            # in the projectmeta API call, and would normally be converted in the
-            # above clean method.)
-            cleaned_data['issuetype'] = {'id': cleaned_data['issuetype']}
-
-        try:
-            response = client.create_issue(cleaned_data)
-        except Exception as e:
-            self.raise_error(e)
-
-        issue_key = response.get('key')
-        if not issue_key:
-            raise IntegrationError('There was an error creating the issue.')
-
-        issue = client.get_issue(issue_key)
-
-        return {
-            'title': issue['fields']['summary'],
-            'description': issue['fields']['description'],
-            'key': issue_key,
-        }
-
-    def sync_assignee_outbound(self, external_issue, user, assign=True, **kwargs):
-        """
-        Propagate a sentry issue's assignee to a jira issue's assignee
-        """
-        client = self.get_client()
-
-        jira_user = None
-        if assign:
-            for ue in user.emails.filter(is_verified=True):
-                try:
-                    res = client.search_users_for_issue(external_issue.key, ue.email)
-                except (ApiUnauthorized, ApiError):
-                    continue
-                try:
-                    jira_user = [
-                        r for r in res if r['emailAddress'] == ue.email
-                    ][0]
-                except IndexError:
-                    pass
-                else:
-                    break
-
-            if jira_user is None:
-                # TODO(jess): do we want to email people about these types of failures?
-                logger.info(
-                    'jira.assignee-not-found',
-                    extra={
-                        'integration_id': external_issue.integration_id,
-                        'user_id': user.id,
-                        'issue_key': external_issue.key,
-                    }
-                )
-                return
-
-        try:
-            client.assign_issue(external_issue.key, jira_user and jira_user['name'])
-        except (ApiUnauthorized, ApiError):
-            # TODO(jess): do we want to email people about these types of failures?
-            logger.info(
-                'jira.failed-to-assign',
-                extra={
-                    'organization_id': external_issue.organization_id,
-                    'integration_id': external_issue.integration_id,
-                    'user_id': user.id,
-                    'issue_key': external_issue.key,
-                }
-            )
-
-    def sync_status_outbound(self, external_issue, is_resolved, project_id, **kwargs):
-        """
-        Propagate a sentry issue's status to a linked issue's status.
-        """
-        client = self.get_client()
-        jira_issue = client.get_issue(external_issue.key)
-        jira_project = jira_issue['fields']['project']
-
-        try:
-            external_project = IntegrationExternalProject.objects.get(
-                external_id=jira_project['id'],
-                organization_integration_id__in=OrganizationIntegration.objects.filter(
-                    organization_id=external_issue.organization_id,
-                    integration_id=external_issue.integration_id,
-                )
-            )
-        except IntegrationExternalProject.DoesNotExist:
-            return
-
-        jira_status = external_project.resolved_status if \
-            is_resolved else external_project.unresolved_status
-
-        # don't bother updating if it's already the status we'd change it to
-        if jira_issue['fields']['status']['id'] == jira_status:
-            return
-
-        transitions = client.get_transitions(external_issue.key)
-
-        try:
-            transition = [
-                t for t in transitions if t['to']['id'] == jira_status
-            ][0]
-        except IndexError:
-            # TODO(jess): Email for failure
-            logger.warning(
-                'jira.status-sync-fail',
-                extra={
-                    'organization_id': external_issue.organization_id,
-                    'integration_id': external_issue.integration_id,
-                    'issue_key': external_issue.key,
-                }
-            )
-            return
-
-        client.transition_issue(external_issue.key, transition['id'])
-
-    def _get_done_statuses(self):
-        client = self.get_client()
-        statuses = client.get_valid_statuses()
-        return {
-            s['id'] for s in statuses if s['statusCategory']['key'] == 'done'
-        }
-
-    def should_unresolve(self, data):
-        done_statuses = self._get_done_statuses()
-        c_from = data['changelog']['from']
-        c_to = data['changelog']['to']
-        return c_from in done_statuses and \
-            c_to not in done_statuses
-
-    def should_resolve(self, data):
-        done_statuses = self._get_done_statuses()
-        c_from = data['changelog']['from']
-        c_to = data['changelog']['to']
-        return c_to in done_statuses and \
-            c_from not in done_statuses
 
 
 class JiraIntegrationProvider(IntegrationProvider):
